Move ConsoleReport to api module - #171
