+++ detekt-cli/src/main/kotlin/io/gitlab/arturbosch/detekt/cli/console/ComplexityReport.kt
@@ -1,6 +1,9 @@
 package io.gitlab.arturbosch.detekt.cli.console
 
+import io.gitlab.arturbosch.detekt.api.ConsoleReport
 import io.gitlab.arturbosch.detekt.api.Detektion
+import io.gitlab.arturbosch.detekt.api.PREFIX
+import io.gitlab.arturbosch.detekt.api.format
 import io.gitlab.arturbosch.detekt.core.COMPLEXITY_KEY
 import io.gitlab.arturbosch.detekt.core.LLOC_KEY
